simplify
- we already have parse catalog script which is for testing
- we already execute that script in fiab-core's just val
- hence we move the newly added functionality in test_artifacts_json in there, to keep responsibility cleanly separate
- we increase the hit surface of the backend's just val, to fire even when install/artifacts.json only is changed
