+++ .github/workflows/ci.yml
@@ -44,7 +44,6 @@ jobs:
       frontend: ${{ steps.filter.outputs.frontend || steps.all-true.outputs.frontend }}
       cli: ${{ steps.filter.outputs.cli || steps.all-true.outputs.cli }}
       cicd_scripts: ${{ steps.filter.outputs.cicd_scripts || steps.all-true.outputs.cicd_scripts }}
-      artifacts: ${{ steps.filter.outputs.artifacts || steps.all-true.outputs.artifacts }}
     steps:
       - uses: actions/checkout@v4
       - uses: dorny/paths-filter@v3
@@ -54,6 +53,7 @@ jobs:
         with:
           filters: |
             backend:
+              - 'install/**'
               - 'backend/**'
               - '.github/workflows/ci.yml'
             frontend:
@@ -65,10 +65,6 @@ jobs:
             cicd_scripts:
               - '.github/**'
               - 'scripts/cicd/**'
-            artifacts:
-              - 'install/artifacts.json'
-              - 'backend/packages/fiab-core/**'
-              - '.github/workflows/ci.yml'
       # For workflow_dispatch and push events, set all outputs to true
       - name: Set outputs for non-PR events
         if: github.event_name != 'pull_request'
@@ -78,7 +74,6 @@ jobs:
           echo "frontend=true" >> "$GITHUB_OUTPUT"
           echo "cli=true" >> "$GITHUB_OUTPUT"
           echo "cicd_scripts=true" >> "$GITHUB_OUTPUT"
-          echo "artifacts=true" >> "$GITHUB_OUTPUT"
 
   macos:
     name: macos
@@ -130,7 +125,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: extractions/setup-just@v3
       - uses: astral-sh/setup-uv@v7
-      - uses: actions/setup-node@v6
+      - uses: actions/setup-node@v6 # TODO do we need the node when we dont build fe?
       - run: |
           set -euo pipefail
           pushd backend
@@ -145,6 +140,7 @@ jobs:
           fi
           popd
 
+          # TODO this does not really cover checking changes in install/*, except for artifacts.json tested by fiab-core -- we want to increase the coverage therin!
           just val
 
   frontend:
@@ -226,17 +222,3 @@ jobs:
       - uses: actions/checkout@v4
       - uses: extractions/setup-just@v3
       - run: just val-cicd-scripts
-
-  artifacts-catalog:
-    name: Artifacts catalog
-    needs: changes
-    # NOTE: Run when the catalog or its schema changes. Hits the artifact store over the network, so
-    # it is deliberately a single job rather than part of the backend test matrix.
-    if: needs.changes.outputs.artifacts == 'true'
-    runs-on: ubuntu-latest
-    timeout-minutes: 10
-    steps:
-      - uses: actions/checkout@v4
-      - uses: extractions/setup-just@v3
-      - run: just val-artifacts
-
