+++ backend/packages/fiab-core/scripts/parse_test_artifacts_catalog.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python3
+#
 # (C) Copyright 2026- ECMWF.
 #
 # This software is licensed under the terms of the Apache Licence Version 2.0
@@ -10,24 +12,17 @@
 """Checks on the shipped artifacts catalog (``install/artifacts.json``).
 
 The url check talks to the artifact store, but only via HEAD -- no checkpoint is downloaded, so it
-costs a handful of round trips and runs by default. What it cannot distinguish on its own is a
-checkpoint that is genuinely gone from a store that is merely unreachable (offline dev machine, store
-outage), so a transport-level failure is a skip. Set ``FIAB_ARTIFACT_STORE_REQUIRED=yes`` -- as the
-ci job guarding catalog changes does -- to turn that skip into a failure.
-"""
+costs a handful of round trips and runs by default."""
 
 import os
 import urllib.error
 import urllib.request
-from collections.abc import Iterator
+from collections.abc import Iterator, Sequence
 from concurrent.futures import ThreadPoolExecutor
 from pathlib import Path
 
-import pytest
-
 from fiab_core.artifacts import ArtifactStoreId, CommonArtifactMetadata, parse_json
 
-CATALOG_PATH = Path(__file__).parents[4] / "install" / "artifacts.json"
 REQUEST_TIMEOUT_S = 15
 # a store outage looks the same as an offline laptop, so one retry before we conclude unreachable
 REQUEST_ATTEMPTS = 2
@@ -35,11 +30,13 @@ REQUEST_ATTEMPTS = 2
 # (status, content_length) on an http response, or (None, reason) if the store could not be reached
 HeadResult = tuple[int | None, int | str | None]
 
+Artifacts = list[tuple[str, CommonArtifactMetadata]]
 
-def _catalog() -> list[tuple[str, CommonArtifactMetadata]]:
-    if not CATALOG_PATH.is_file():
-        pytest.skip(f"catalog not present at {CATALOG_PATH} -- running outside a source checkout")
-    parsed = parse_json(ArtifactStoreId("validation"), CATALOG_PATH.read_text(), lambda _common, _specific: (True, None))
+
+def read_and_parse(pth: Path) -> Artifacts:
+    if not pth.is_file():
+        raise ValueError(f"catalog not present at {pth} -- running outside a source checkout")
+    parsed = parse_json(ArtifactStoreId("validation"), pth.read_text(), lambda _common, _specific: (True, None))
     return [(composite_id.artifact_local_id, resolved.common) for composite_id, resolved in parsed]
 
 
@@ -60,20 +57,13 @@ def _head(url: str) -> HeadResult:
     return None, reason
 
 
-def test_catalog_parses() -> None:
-    """The shipped catalog must validate against the current schema."""
-    artifacts = _catalog()
-    assert artifacts, "catalog contains no artifacts"
-
-
-def test_catalog_urls() -> None:
+def validate_catalog_urls(artifacts: Artifacts) -> None:
     """Every catalog url must resolve, and disk_size_bytes must match the actual download size.
 
     Guards against the two ways the catalog silently rots: a checkpoint being moved or renamed in
     the artifact store, and a disk_size_bytes copy-pasted from a sibling entry (it is display-only,
     so a wrong value is invisible until a user reads it).
     """
-    artifacts = _catalog()
 
     # several entries legitimately share one checkpoint (eg the aarch64 variant), so fetch each url once
     by_url: dict[str, list[str]] = {}
@@ -87,9 +77,7 @@ def test_catalog_urls() -> None:
     unreachable = [f"{url}: {reason}" for url, (status, reason) in results.items() if status is None]
     if unreachable:
         message = "could not reach the artifact store:\n" + "\n".join(unreachable)
-        if os.environ.get("FIAB_ARTIFACT_STORE_REQUIRED", "no") == "yes":
-            pytest.fail(message)
-        pytest.skip(f"{message}\nset FIAB_ARTIFACT_STORE_REQUIRED=yes to treat this as a failure")
+        raise ValueError(message)
 
     def problems() -> Iterator[str]:
         for url in urls:
@@ -103,4 +91,25 @@ def test_catalog_urls() -> None:
                 yield f"{local_id}: disk_size_bytes is {common.disk_size_bytes} but {common.url} is {content_length} bytes"
 
     found = list(problems())
-    assert not found, "artifacts catalog is out of sync with the artifact store:\n" + "\n".join(found)
+    if found:
+        message = "artifacts catalog is out of sync with the artifact store:\n" + "\n".join(found)
+        raise ValueError(message)
+
+
+def main(argv: Sequence[str] | None = None) -> None:
+    import argparse
+
+    parser = argparse.ArgumentParser()
+    parser.add_argument("path", type=Path)
+    args = parser.parse_args(argv)
+    pth = args.path
+
+    # 1. read and parse
+    artifacts = read_and_parse(pth)
+
+    # 2. validate individual urls
+    validate_catalog_urls(artifacts)
+
+
+if __name__ == "__main__":
+    main()
